+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.2] - 2020-12-25
+
+### Fixed
+
+- Type error when trying to serialize and deserialize object with null values.
+- Explicit deserialization of properties.
+
 ## [1.2.1] - 2020-12-24
 
 ### Fixed